Update Appendix A_online_supp.R
Line 150: I suggest to change the code if criteria from v.p.it  == 1 to all.equal(sum(v.p.it),1) for exact number comparisons. I adapted the code with my own example and got the error "probabilities do not sum to 1" even though they did. The suggested change did not give an error and does not change the logic of the code. It apparently has something to do with the tolerance level when comparing numbers with "==".